🍗 :: exporting method
